wxwidgets: insert more debugging output into code
This time we use

+#define PLMemoryMap_DEBUG
+//#undef PLMemoryMap_DEBUG

in drivers/wxwidgets_comms.h to keep all this debugging noise in the output until we solve this.  But for the
release must remember to reverse the sense of the commenting out done
above.

Note this debug output shows what naively appears to me (because there may be an
easy explanation I just don't understand due to lack of C++ skills) to be
a deadloock of some kind which I will discuss on list.