qt device driver: drop use of automoc
This was an example of automoc overkill. All the relevant
automoc-generated symbols are already available and exported from the
plplotqt library which the qt shared object links to.  So it was
redundant to regenerate those symbols again with automoc for the qt
shared object, and although the result worked on Linux it appeared to
be associated with build problems on Cygwin.  So my hope is that by
removing this redundancy (which also works fine on Linux, see below),
the Cygwin problems associated with the build of qt will disappear.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by building the test_c_qtwidget target with no
obvious configure-, build-, or run-time issues.

Furthermore,

ldd -r drivers/qt.so |& less ,

nm --defined-only --extern-only --demangle drivers/qt.so ,

and

nm --undefined-only --demangle drivers/qt.so |less

all gave the expected (good) results.

N.B. testing on Cygwin is requested to make sure this fix and a prior
similar fix for pyqt4 work well on that platform.