Fixed problem with rotated wxWidgets text
When text was rotated and there was a style change mid line the text was
rendered in the wrong position. Basically for a 90 degree rotatio the x
and y translations relative to the text origin were reversed.

A similar but more subtle issue was that text with newlines was also not
aligned properly as it was assumed that the scale for converting between
plplot units and pixels was the same which only occured when the device
context was square.

These problems were fixed by passing an origin through the function
calls and passing coordinates relative to that origin. This meant that
the relative positions got correctly rotated.