wxwidgets binding: insert debug messages
Add debug messages that show sequence of events for a bug that happens
on some but not all Linux distributions with wxwidgets3.0/3.1.  The
cause is that frame->Create does not always trigger an OnCreate()
event on time before the Plot() call.

Tested by: Pedro Vicente <pedro.vicente@rhw9121.star1.nesdis.noaa.gov>
on CentOS Linux with wxwidgets3.1 built from source with the following
results from building the test_wxPLplotDemo target:

11:09:13: Debug: wxPLplotwindow::wxPLplotwindow
11:09:13: Debug: frame->Create
11:09:13: Debug: pls NULL
11:09:13: Debug: wxPLplotwindow::OnCreate

Clearly this demonstrates the bug with the OnCreate() event happening
too late.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) with the system version of wxwidgets-3.0.2 with
the following results from building the test_wxPLplotDemo target:

23:09:34: Debug: wxPLplotwindow::wxPLplotwindow
23:09:34: Debug: frame->Create
23:09:34: Debug: wxPLplotwindow::OnCreate
23:09:34: Debug: Plot()

Clearly in this case the OnCreate() event happens on time in contrast
with Pedro's result.  This demonstrates the bug is platform-dependent
where because timing might be involved, platform should be interpreted
as generally as possible including the hardware/OS version as well as
wxwidgets and gtk+ software.  The above results were without
VirtualBox, but in addition all of Pedro's tests failed on a
VirtualBox setup he was using where host was a Windows OS and the
guests were various Linux distributions with system versions of
wxwidgets. Pedro also got a bad test for a non-VirtualBox Ubuntu
distro with wxwidgets was built from source while Phil got a good test
for a non-VirtualBox Ubuntu distro for system wxwidgets.  So there are
no clear patterns here.