wxwidgets binding: workaround fix for delayed OnCreate event
Add a workaround fix for cases where frame->Create does not trigger an
OnCreate() event before the Plot() in test_wxPLplotDemo. The solution
is to create the stream in a overidden Show() method, that is always
called on window creation.

Tested by: Pedro Vicente <pedro.vicente@rhw9121.star1.nesdis.noaa.gov>
on CentOS Linux with wxwidgets3.1 built from source with the following
results from building the test_wxPLplotDemo target:

13:16:18: Debug: wxPLplotwindow::wxPLplotwindow
13:16:18: Debug: frame->Create
13:16:18: Debug: wxPLplotwindow::Show
13:16:18: Debug: wxPLplotwindow::CreateStream
13:16:19: Debug: Plot()
13:16:19: Debug: wxPLplotwindow::OnCreate
13:16:19: Debug: wxPLplotwindow::CreateStream

The above debug sequence shows that the stream is
valid before the plot is made.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) with the system version of wxwidgets-3.0.2 with
the following results from building the test_wxPLplotDemo target:

23:41:12: Debug: wxPLplotwindow::wxPLplotwindow
23:41:12: Debug: frame->Create
23:41:12: Debug: wxPLplotwindow::Show
23:41:12: Debug: wxPLplotwindow::CreateStream
23:41:12: Debug: wxPLplotwindow::OnCreate
23:41:12: Debug: wxPLplotwindow::CreateStream
23:41:12: Debug: Plot()

In this case, Plot() occurs in a different order but the net effect is
the same as Pedro's case; an OnCreate() event occurs before Plot().
Furthermore, the plot result on this system looks the same as it
did before this fix (as expected from the above event sequence).