Update release notes
These preliminary release notes have not been finalized yet because
the preliminary ChangeLog of commit messages for this release has only
been reviewed for topics to include in these release notes up to the
first NAGFOR-related commit.  Also, the section on the documentation
changes is forward looking (i.e., stores on-going notes of a
work-in-progress, my documentation work on a private topic branch that
has not been pushed yet).