Upgrade debug output for wxwidgets using the PLPLOT_WX_DEBUG_OUTPUT and PLPLOT_WX_NANOSEC CMake options.
Implement the option PLPLOT_WX_DEBUG_OUTPUT which defaults to OFF.
However, when the user specifies ON, then
debug output is generated for wxwidgets device driver, binding, and example
wherever the developers have used the macro

    PLPLOT_wxLogDebug(<some string>);

in the code.

N.B. When the default OFF option is used, that macro is #define'd but empty, and
therefore there is no code that is compiled from the above use of
the macro and therefore no wxwidgets debug output for this case.

Implement the option PLPLOT_WX_NANOSEC which is only meaningful if
PLPLOT_WX_DEBUG_OUTPUT is ON.  If both these options are ON, then a
nanosec time stamp is applied to the debug message that is output.
However, PLPLOT_WX_NANOSEC defaults to OFF (so that nanosec time stamp
is not applied) because this option is completely experimental, i.e.,
it will introduce build errors if the user's system does not properly
support the clock_gettime function which is a POSIX option so
clock_gettime is mostly available only on Linux.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) for 3 configuration cases with the following cmake command-line
options:

(1) -DPLPLOT_WX_DEBUG_OUTPUT=ON -DPLPLOT_WX_NANOSEC=ON
(2) -DPLPLOT_WX_DEBUG_OUTPUT=ON
(3) neither of these options

The result for (1) when building the test_wxPLplotDemo target was

19:00:38: Debug: nanosecs since epoch = 1808113138508376: wxPLplotwindow::wxPLplotwindow
19:00:38: Debug: nanosecs since epoch = 1808113147285531: frame->Create
19:00:38: Debug: nanosecs since epoch = 1808113151908731: wxPLplotwindow::Show
19:00:38: Debug: nanosecs since epoch = 1808113151958780: wxPLplotwindow::CreateStream
19:00:38: Debug: nanosecs since epoch = 1808113172019557: wxPLplotwindow::OnCreate
19:00:38: Debug: nanosecs since epoch = 1808113172092596: wxPLplotwindow::CreateStream
19:00:38: Debug: nanosecs since epoch = 1808113181701454: Plot()

The result for (2) when building the test_wxPLplotDemo target was

20:38:39: Debug: wxPLplotwindow::wxPLplotwindow
20:38:39: Debug: frame->Create
20:38:39: Debug: wxPLplotwindow::Show
20:38:39: Debug: wxPLplotwindow::CreateStream
20:38:39: Debug: wxPLplotwindow::OnCreate
20:38:39: Debug: wxPLplotwindow::CreateStream
20:38:39: Debug: Plot()

And the result for (3) when building the test_wxPLplotDemo target was
no debug output at all.

I find these wxwidgets debug output results (i.e., no debug output at
all by default) completely satisfactory for the release, although I
will likely be deploying additional usage of

PLPLOT_wxLogDebug(<some string>);

to try and track down the long pause bug.