Added a function to check if the window is ready to accept plot commands.
This function calls GetStream() to get a pointer to the wxPLplotstream
and checks if this pointer is NULL as a method for checking if
everything is ready to accept plot commands. This method is then used
internally in a few locations to check if things are ready in OnMouse
and OnPaint. Performing these checks implies that it is possible for a
paint or mouse event to be generated before a create event. I'm not sure
if this is the case, but the wxWidgets docs are a bit vague so there is
no harm in using the checks.