Fix bug with plotting in wxPLplotDemo
On most systems it seems that when wxWindow::Create() is called that
triggers a create event, but within the Create() function the event loop
is checked and the create event is processed immediately and OnCreate()
is called setting up the wxPLplotstream. This means we can begin using
the wxPLplotstream immediately after the call to Create().

However on some systems the create event is not processed immediately.
This has been observed on some Linux systems but apparently
identical systems seem to have different behaviour. The cause of this is
not clear.

To deal with this we now check the wxPLplotwindow is ready and wait for
it if not, yielding to allow the create event to be processed. Another
option might be to catch idle events and each time check the window is
ready then do the plotting at this point. However this is a more complex
proceedure and it would partly obfuscate what should be a simple
example.