Use /dev/urandom in preference to /dev/random in wxWidgets driver.
The wxWidgets driver needs to create a random name for a memory map to
avoid name clashes with concurrent instances. The seed for the random
number generator was created from /dev/random, however this blocks and
caused up to a 1-15 second delay in some instances while waiting for
entropy. Swap to using /dev/urandom when possible which is not as
random, but suits our needs and does not block.