Update the release notes for 5.12.0
N.B. The ChangeLog consisting of all the commit messages for the ~400
commits that are included in this release has been reviewed to help
insure that these release notes mention all the major PLplot
developments that have occurred for this release.  So virtually all
sections of these release notes should be ready for release.  However,
my work on the large documentation update is still in progress so that
section of these release notes will likely require further revision.