Add convenience batch file for running the examples under bare Windows
Tested by: Arjen Markus (arjenmarkus@users.sourceforge.net)

Due to the lack of sufficiently popwerful tools in the Windows environment (the Windows powershell might provide the tools that are required, but I do not know enough about it for an implementation) the automated testing on this platform is limited. To help with this I have written a small batch file and a dedicated comparison program.

The idea is simple: run the examples for the various programming languages and compare the output files. It presumes that the examples have been built already, so it does not test anything else than the build that is present.

To compare the files I wrote a small dedicated C program - experiments with the diff command from MinGW-w64/MSYS2 failed, as it appears you cannot run such programs outside the MinGW-w64/MSYS2 shell (merely setting the path is not enough).

To use it: go to the top-most directory holding the built examples (that is, the directory with subdirectories like bindings, examples, dll, ...) and run the script "run_examples_windows.bat". That is it. Redirect the output for later inspection.

Note: it is a fairly simple script - it assumes that all executable programs whose name start with an "x" are PLplot examples with C equivalents. Some experimental programs are present, however, for instance x01cc.exe, so that the comparison program fails to compare the right files. Ignore the resulting false messages.