Fix infinite yield issue in wxPLplotDemo
It turns out that on Linux (I think all systems, not just with remote
X11 servers) the wxEVT_CREATE event is not triggered until after Show()
is called. This was causing an infinite loop while the code repeatedly
called Yield to wait for this event before calling Show().

Provided two separate windows in the example now showing two ways to do
this properly. One is using inheritence and having a class which catches
wxEVT_IDLE events, checks if the window is ready and then does the
plotting. The second does not use inheritance, but just has to make sure
that Show() is called, followed by some waiting and Yielding (probably
only needed for remote X11 servers) before we attempt to do any
plotting.