Fixed text transform issues in wxWidgets when not using wxGraphicsContext transforms
The wxWidgets driver uses a number of systems to transform text. Some
wxDCs support transforms, if so these were used. Up until recently
wxGCDC did not support transforms, but the underlying wxGraphicsContext
does, so these were used where needed. If neither of these options was
possible then we just do simple translate and rotate transforms.

Unfortunately ony the wxGraphicsContext code was regularly checked as
wxGCDCs are used by wxPLViewer and much other display code.

In the very recent past (during wxWidgets 3.1 development) wxGCDC
started supporting wxDC transformations and the plplot driver logic was
arrange such that they were used. This exposed a bug in the use of those
transforms. Further investigation revealed a bug in the simple translate
rotate code, including a logic bug that meant this code was only used
pre wxWidgets 29.0.2 and a syntax error that meant this code actually
didn't compile pre wxWidgets 29.0.2!

This commit fixes the logic and compilation bugs that existed in the use
of the transformations.

It also reorders the if statements and the way that the code checks for
a wxGCDC. This makes it clearer that in the case where wxGCDCs now
support wxDC transformations it is these transformations that will be
used.

It should also be noted that at some point if we bump the minimum
wxWidgets version to 3.2 (once this version gets released) then we
should be able to ditch all wxGCDC specific code.

Code tested on Windows.