DocBook documentation: Update instructions for documentation developers
The section concerning test results was changed in recognition of the
fact that konsole (the KDE terminal emulator that uses client-side
system fonts where glyphs are selected by font-config) yields
essentially perfect UTF-8 test results for info pages.  This is a huge
improvement compared to the missing Mandarin, Korean, and Devanagari
(Hindi script) glyphs and the glyph rendering in the wrong order for some
CTL languages that you get with xterm (the X terminal emulator that
uses server-side X fonts without the benefit of font-config).