DocBook documentation: Remove detailed API discussions from the "Simple Use of PLplot" chapter
Detailed discussion of the API is the job of the "The Common API for
PLplot" chapter.  So repeating this information in other chapters is
unnecessary and leads to maintenance issues.  Instead, a simple
cross-reference to the appropriate API is the correct thing to do in
other chapters, and this commit implements that idea for the
"Simple Use of PLplot" chapter.