Build system: temporarily remove the check_all target dependency on check_plplot_h.inc
The issue is that check_plplot_h.inc now encounters run-time issues when attempting
to generate its check file using ocaml.  This is likely due to a namespace
or installation issue with ocaml.  Until this issue is fixed, remove the
check_all dependency on check_plplot_h.inc so that check_all will continue to work.