Build system: implement and use PREBUILT_DOX_DOC option
This option plays the same role for doxygen documentation as the
PREBUILT_DOC option plays for the DocBook documentation.

Those using a tarball with prebuilt doxygen documentation should use
this option to allow use of that documentation (e.g., installation of
it) just as if it was built as part of the build process.  So, for
example, we use this option when testing installation of a release tarball
in the make_tarball.sh script.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by running

scripts/make_tarball.sh -c -i /tmp/plplot_install

and checking the doxygen build when creating the tarball
and checking the doxygen install from the prebuilt
doxygen documentation in that tarball.