Release process: change when cumulated release notes are stored.
Those cumulated release notes were previously stored in
OLD-README.release but now that file is renamed as
README.cumulated_release, and the final README.release notes are
prepended to that file as the last step before the release is tagged.
Therefore, README.cumulated_release should be up to date for the
release tag and also the release tarball.  I also added
README.cumulated_release to the files that are installed.