The cairo device driver: fix bug in aStream->xdrawable_mode logic
The test for aStream->xdrawable_mode had to be moved from
plD_eop_xcairo (where it effectively did nothing) to
the start of plD_wait_xcairo where it returns immediately
if aStream->xdrawable_mode is true.

aStream->xdrawable_mode is only true if external_drawable is true
and that value is only set by the extXdrawable_demo application.
So the net effect of this bug fix is essentially all of
plD_wait_xcairo is skipped for the extXdrawable_demo application
case which is consistent with all of that code being skipped
in like case when it was part of plD_eop_xcairo.

As a result of this change the run-time error for the
test_extXdrawable has been addressed (see below) so
I have reinstated test_extXdrawable as a dependency of generic
test targets again.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian testing) by building the test_c_xcairo, test_extcairo, and
test_extXdrawable targets.  The first two are unaffected
by this change (i.e., they continue to work fine as expected), and in addition
the run-time error when running extXdrawable_demo for the 3rd target has
now been addressed (because the plD_wait_xcairo code is skipped as it should be for
this case).