wxwidgets IPC: Use a struct for the shared memory area
If the PL_HAVE_UNNAMED_POSIX_SEMAPHORES macro is #defined
use a shmbuf struct for the shared memory buffer that contains
a fixed-size character array buf and some other essential
fields.

This change potentially allows (a) (with the cnt element of the shmbuf
struct) keeping track of how many bytes are used in the char array,
and (b) (with the wsem and rsem elemenst of the shmbuf struct) use of
two unnamed semaphores to control the IPC.  But for now cnt, wsem, and
rsem remain unused.

If the PL_HAVE_UNNAMED_POSIX_SEMAPHORES macro is #defined
this change means that the getBuffer and getSize methods
of the PLMemoryMap class respectively return the location
and size of the character array buf within the shmbuf struct.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by building test_c_wxwidgets (to build all prerequisites)
and then running various C examples with -dev wxwidgets without
any noticeable run-time issues.  Also, I took a look in /dev/shm
to determine the size of the memory buffer which is now 72 + 1024*1024
rather than 1024*1024 as before.  The three extra fields in shmbuf
are of types sem_t, sem_t, and size_t with corresponding sizes of
32 (according to /usr/include/x86_64-linux-gnu/bits/semaphore.h), 32, and 8
which is exactly consistent with the 72 if there are no alignment bytes
in the struct.