Initialised remaining member variables in PlDevice
although Alan had good valgrind results. It looked to me like if someone
happened to try to draw unicode character 0x0000 for their first
string then the current check for the size of the previous character
could succeed and return garbage. Plus it is good practiace to initilise
memer variables to sensible initial values.

All these variables are now initialised to 0. Which means if the first
size requested is of a 0x0000 character with size 0 and FCI 0 then the
check for an existing size will succeed and will return sizes of 0,
which is consistent.