wxwidgets IPC: implement use of unnamed semaphores
This commit completes case 1 that was discussed on list where for
the present one-semaphore approach (which requires use of sleep and
mutex API) an unnamed semaphore is used (if
-DPL_HAVE_UNNAMED_POSIX_SEMAPHORES=ON) instead of a named semaphore.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by building the target test_c_wxwidgets and
subsequently running

examples/c/x16c -dev wxwidgets

For both the -DPL_HAVE_UNNAMED_POSIX_SEMAPHORES=ON and
-DPL_HAVE_UNNAMED_POSIX_SEMAPHORES=OFF cases.  As expected (since the
same fundamental single-semaphore approach is used in both cases) I
could find no difference in results between those two options.  For
example, for both cases the time when wxPLViewer GUI's were on the
screen was 60 seconds for test_c_wxwidgets and 14 seconds for the
example 16 test above (where the enter key was pressed a lot to keep
the page display moving as fast as possible).