wxwidgets IPC: implement a lot more of the wxwidgets two-semaphore IPC approach
For the case when -DPL_WXWIDGETS_IPC2=ON the IPC is now implmeented as
follows:

* The PLTwoSemaphores class has been matured with the addition of a
  number of methods (such as initializeToValid, initializeToInvalid,
  areBothSemaphoresValid, and areBothSemaphoresBlocked) that have
  proved useful for IPC using the two-semaphores approach.
  Furthermore, The alternative constructor has been dropped, and the
  destructor has been simplified.

* Add transmissionType and plbufAmountToTransmit fields to the
  MemoryMapHeader struct rather than having such header-type fields
  mixed with plbuf information in the data transmitted back and forth
  via IPC between writer (-dev wxwidgets) and reader (wxPLViewer).

* Separate header data from plbuf data in the data that are
  transmitted between writer and reader.  This means that within the
  shmbuf struct that MemoryMapHeader header data are transmitted via
  the header field, and plbuf data are transmitted via the data field.

* Separate header data (as a private MemoryMapHeader m_header variable
  within the wxPLDevice class) from shmbuf data (which is strictly
  used just to transmit the header data and not store it).  The same
  change has implemented for the wxPlFrame class.  The net result of
  these changes is a complete separation of instance header
  information from shmbuf header information.

* Change all use of memmove to memcpy (since we now have complete
  separation of storage of header information (i.e., in the private
  instance m_header variables) from the staging area in shmbuf used to
  tranfer that header information.

* For the PLMemoryMap class implement the moveBytesWriter,
  moveBytesReader, moveBytesReaderReversed, and
  moveBytesWriterReversed methods where the first two implement
  transmission from writer to reader and the second two implement
  transmission from reader to writer.

For the case when -DPL_WXWIDGETS_IPC2=ON, the current status is this
new IPC method is used in a mature way on the writer (-dev wxwidgets)
side but so far is mostly just used to confirm the data transmissions
completed successfully on the reader (wxPLViewer) side.  That is, at
run time the new IPC method is used to transfer header data (including
viewerOpenFlag = 1) from reader (wxPLViewer) to writer (-dev
wxwidgets) at the start, use wxPLDevice::TransmitBuffer on the writer
side as appropriate (including sending the header completeFlag = 1 at
the end) to send a header and optional plbuf data to the reader, and
on the reader side reading that information with
wxPlFrame::ReadTransmission, dumping out some key header values (to
show the data were received), but not acting on that data except to
exit when header completeFlag = 1 is received at the end.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by building the wxwidgets and wxPLViewer targets for
the case when -DPL_WXWIDGETS_IPC2=ON, and run-time testing the
result using

examples/c/x08c -dev wxwidget

The result shows correct transmission of the header viewerOpenFlag = 1
from reader to writer, correct transmission of BOP, plbuf data, and
EOP trom writer to reader (once per page), and correct transmission of header
completeFlag = 1 from writer to reader before the reader exits.  And
all these plbuf data (roughly 2MBytes for the 10 pages of that
example) are assembled and transferred in less than a second so it
appears the transmission speed is already comparable with the
-DPL_WXWIDGETS_IPC2=OFF case.

For the -DPL_WXWIDGETS_IPC2=OFF case, the test_c_wxwidgets target built
without issues and

examples/c/x08c -dev wxwidgets

(note no -np option) executed with no issues.

So it appears the large code changes for the -DPL_WXWIDGETS_IPC2=ON
case are not affecting the -DPL_WXWIDGETS_IPC2=OFF case.