wxwidgets two-semaphore IPC: use consistent transmission types
Change the type of transmissionType in the MemoryMapHeader struct from
size_t to unsigned char (and reorder the elements of that struct to
reduce padding for this case) to be consistent with the unsigned char
type of the possible transmission flags constants that are compared
with transmissionType.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by rebuilding for the -DPL_WXWIDGETS_IPC2=ON case
the wxwidgets, wxPLViewer, and x09c targets and running

examples/c/x09c -dev wxwidgets

without issues.