wxwidgets two-semaphore IPC: Attempt to avoid race when data flow reverses
When data flow is all in the same direction (e.g., moveBytesWriter
calls followed consecutively by additional such calls on the writer
(-dev wxwidgets) side) each moveBytesWriter call sets up the next with
no chance for races.  And the same is true of consecutive
moveBytesReaderReversed calls on the reader (wxPLViewer) side.
However, a call to moveBytesReaderReversed on the reader side followed
by a call to moveBytesWriter on the writer side (i.e., where the
direction of data flow is reversed) creates a race because there is no
guarantee (although it should happen virtually all the time) the OS
scheduler will allow moveBytesReaderReversed to finish before
moveBytesWriter is started.  And similarly there is an equivalent race
when a call to moveBytesWriter on the writer side is followed by a
call to moveBytesReaderReversed on the reader side.

This commit attempts to avoid those races by inserting (whenever there
is a reverse in the flow of data) a 10 ms sleep.  This should not be
an efficiency concern in the slightest because the sleep is only
executed when the direction of data flow is reversed, and that only
occurs for an initialization handshake now and interactive use of -dev
wxwidgets (later when interactive capabilities are implemented for the
-DPL_WXWIDGETS_IPC2=ON case).

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by rebuilding for the -DPL_WXWIDGETS_IPC2=ON case
the wxwidgets, wxPLViewer, and x09c targets and running

examples/c/x09c -dev wxwidgets

without issues.

N.B. my current thinking is the above 10 ms sleep is way more than
enough.  The score is currently that it has worked once in the above
test, but only time will tell whether the
"PLMemoryMap::moveBytesWriter: attempt to start transfer with
semaphores not in correct blocked state (likely due to a race)."
exception or equivalent exception for
PLMemoryMap::moveBytesReaderReversed is ever thrown again.