wxwidgets three-semaphore IPC: first attempt at implementing interactivity
Here is what should be happening as a result of this
attempted implementation:

* In example 1 using the -locate option after all normal PLplot calls
to generate the plot have been called (which fills up plbuf with 24272
bytes of information, see below), the locate mode loop generates a
series of plGetCursor calls.

* Each of those plGetCursor calls ends up (for -dev wxwidgets) as a call to
wxPLDevice::Locate which has the following code:

TransmitBuffer( pls, transmissionLocate );
m_outputMemoryMap.receiveBytes( true, &m_header, sizeof ( MemoryMapHeader ) );
*graphicsIn = m_header.graphicsIn;

* TransmitBuffer calls transmitBytes to send header information
(including transmissionLocate = 4 and  plbufAmountToTransmit = 24272
bytes in plbuf to be received by wxPlFrame::ReadTransmission (which is called
by wxPlFrame::OnCheckTimer on a timer event loop) using receiveBytes.

The results are

Before transmitBytes
transmissionType = 4
plbufAmountToTransmit = 24272
viewerOpenFlag = 1
locateModeFlag = 1
completeFlag = 0
After receiveBytes
transmissionType = 4
plbufAmountToTransmit = 24272
viewerOpenFlag = 1
locateModeFlag = 1
completeFlag = 0
Successful read of plbuf

i.e., that transmission of the plbuf information and
transmissionLocate to wxPlFrame::ReadTransmission was a complete success.

* Because plbufAmountToTransmit is non-zero, wxPLViewer appends that
plbuf information to the correct element of the m_pageBuffers array
and conditionally (although the result is the condition is true
because of the above 24272 says there is a lot more data to plot)
calls SetPageAndUpdate as per normal which should update all the
required plotbuf information and configure all the plotbuf data to be
plotted using the Refresh method.

* After plbuf is processed, then wxPlFrame::ReadTransmission
for the transmissionLocate case calls
SetPageAndUpdate one more time (which should have no effect because
the new plbuf information has already been completely processed by the
previous call to SetPageAndUpdate), then sets m_locateMode = true
before exiting wxPlFrame::ReadTransmission and wxPlFrame::OnCheckTimer
to complete processing of that timer event.

* After that exit from wxPlFrame::OnCheckTimer the plot should be
automatically plotted (because of the above call to the Refresh
method), but for some reason (likely a bug) that does not happen.

* Then once a mouse click causes another event, wxPlFrame::OnMouse
collects information about that click into the m_header.graphicsIn
struct, and then the whole header is sent back to -dev wxwidgets with
transmit_Bytes (which because of the three-semaphore design will wait
for the end of the previous transmit_Bytes before proceeding so I
don't think we are up against any race condition here) where
wxPLDevice::Locate receives it with receiveBytes (see line of code
above), and passes it back to the example 1 locate loop which prints
out values collected for a particular mouse click, then (unless there
is an exit from locate mode by a mouse click off of a viewport keeps
repeating the whole process indefinitely.

The above is the theory, but in practice the plot never shows and everything
just hangs.  So there is likely some small bug in the implementation, but
I haven't found it yet.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by building wxwidgets, wxPLViewer, and x01c for the
-DPL_WXWIDGETS_IPC2=ON case and running

examples/c/x01c -dev wxwidgets

with complete success, but

examples/c/x01c -dev wxwidgets -locate

produces the above-mentioned hang with no plot being displayed yet
also the above evidence that all the data was transmitted from -dev
wxwidgets to wxPLViewer without issues (except possibly that
plbufAmountToTransmit = 24272 rather than the value of 24270 that
occurred for the above case without the -locate option).  So it
currently appears that all you need to do to get this fixed is display
the plot after wxPlFrame::OnCheckTimer so that any mouse click after
that on that display will end up as a call to wxPlFrame::OnMouse.

N.B. this issue of the plot not being displayed also occurs for
the -DPL_WXWIDGETS_IPC2=OFF case.  So presumably the same fix
with fix both the -DPL_WXWIDGETS_IPC2=ON and OFF cases.