README.developers: updated git availability and git command-line paragraphs
With regard to git availability, I have now added Xcode and Fink as
sources of git packages on Mac OS X and "Git for Windows", Cygwin, and
MinGW-w64/MSYS2 as sources of git packages on Windows.  I have removed
the msysGit project as a source for git packages on Windows since that
project has now been superseded by "Git for Windows".

With regard to the git command-line, I have improved the justification
for using that rather than some non-standard git GUI, and I have added
more information about how to obtain access to the git command line
for both the Xcode and "Git for Windows" versions of git.