Build system: implement try_run test to determine PL_HAVE_UNNAMED_POSIX_SEMAPHORES
On Windows platforms (i.e., those where the CMake variable WIN32 is
true) the result of this test is the CMake variable
PL_HAVE_UNNAMED_POSIX_SEMAPHORES is undefined and the equivalent C
macro is accordingly not #defined).  On non-Windows platforms (i.e.,
POSIX platforms), a try_run test is run to check for the build and
run-time success of sem_init and sem_destroy with CMake variable
PL_HAVE_UNNAMED_POSIX_SEMAPHORES and equivalent C macro set
accordingly.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by running cmake with the -DPL_WXWIDGETS_IPC3=ON option
and looking at the (correct) results for PL_HAVE_UNNAMED_POSIX_SEMAPHORES.

N.B. a platform concern is the -pthread linking option used for the
try_run test.  Other linking options might have to be used on other
platforms to obtain linking success for the sem_init and sem_destroy
calls used in this test.  However, if this test fails on some
platforms because of this issue, it is only a minor concern since the
fallback of using named semaphores on such platforms should be a good
alternative.