wxwidgets 3-semaphore IPC: tuned PL_SHARED_ARRAY_SIZE to 10*1024
I measured the total IPC time (roughly 18 seconds on my system) taken
to execute all our C standard examples (except for 17 and 20) with
-dev wxwidgets as follows:

* I locally modified utils/wxplframe.cpp to #define PLPLOT_WX_NOPLOT
  (to stop the rendering inefficiency of wxPLViewer from interfering
  with the timing test).

* I built all prerequisites using the test_c_wxwidgets target at first
  (which builds the wxwidgets and wxPLViewer targets as well as all the
  C examples).  Later I just rebuilt the wxwidgets and wxPLViewer targets
  for each change in PL_SHARED_ARRAY_SIZE.

* For each of those builds, I measured the total IPC time using

time(for N in $(seq --format='%02.0f' 0 34|grep -vE '17|20'); do examples/c/x${N}c -dev wxwidgets >&/dev/null; killall -9 wxPLViewer; done)

where the killall command was required to get rid of the empty wxPLViewer GUI left by running
the standard example with PLPLOT_WX_NOPLOT #defined.

By repeating the above test for the same value of
PL_SHARED_ARRAY_SIZE, I discovered that the repeatibility of the total
time results was typically within 2 per cent or less.  Comparison of
PL_SHARED_ARRAY_SIZE of 128 versus 1024 showed a barely significant 4 per cent
improvement in speed, but there was no significant improvement beyond
that for any value of PL_SHARED_ARRAY_SIZE tried above 1024 including
1024*1024.  Therefore, the final PL_SHARED_ARRAY_SIZE adopted for this
commit which is 10*1024 is very likely gross overkill, but still it is
one hundred times smaller than the 1024*1024 size of the circular
buffer used for the -DPL_WXWIDGETS_IPC3=OFF case.

Normally, this large reduction in the amount of shared memory size
that is consumed by -dev wxwidgets and wxPLViewer IPC would not be a
significant advantage on systems with substantial amounts of shared
memory available and no shared memory leaks associated with the
wxwidgets IPC.  However, it is better style to use only the amount of
shared memory that is necessary because on some systems that is a
rather limited resource.  Furthermore, for some unknown reason we
still have a shared memory leak (i.e., when you ctrl-C out of an
example, /dev/shm shows shared memory is "permanently" consumed until
the user logs off the system or executes a rm -f /dev/shm/*plplot*
command).  Therefore, so long as this shared memory leak persists with
wxwidgets IPC the much smaller shared memory buffer that results from
this commit could be an advantage if the user is in the habit of
exiting from -dev wxwidgets in a way (e.g., such as ctrl-C) that
triggers the shared memory leak.