Developer documentation: rework recommended gcc/g++/gfortran compiler options
The principal change is the prior suggestions for the gfortran case
had to be changed to be consistent with our new Fortran binding.  In
particular the -std=f???? -pedantic options had to be dropped because
they created what I consider to be spurious build errors.  This means
we cannot check for Fortran 2003 compliance with gfortran and must
rely on a different Fortran compiler (NAG) to do that.