CMake-based build system for installed examples: Fix bug when ENABLE_CXX is OFF
C++ is a special language case in that we enable that language in the
CMake-build system used for the build tree if at all possible
regardless of the value of ENABLE_cxx.  For ENABLE_cxx OFF, this
enabling of the C++ language allows, e.g., libplplot to link properly
with the C++ compiler when it contains C++ object modules (e.g., the
compiled code for the psttf device driver when that is resident in the
libplplot library).  However, previously our CMake-based build system
for the installed examples did not follow this special C++ enabling
logic so the result was linking errors for any example that linked to
libplplot (when that library contained C++ code) when ENABLE_cxx was
OFF.  This fix addresses that issue.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by running

scripts/comprehensive_test.sh --cmake_added_options "-DDEFAULT_NO_BINDINGS=ON -DENABLE_fortran=ON" --do_test_interactive no

Before the fix, there were libLasi linking errors for the static case
when building the installed examples with the CMake-based build system
for those examples.  (I don't quite understand how I avoided the
equivalent error for the nondynamic case, but perhaps linking is not
checked so rigourously for that shared library case.)  After the fix,
the above script run finished without issues.