Drop "module" from basenames for _plplotc and plplot_widget Python extension modules
The "module" part of basenames for Python extension modules is
optional so we removed it completely.  This change simplifies our
build system from the previous mixed basename case where sometimes the
basename included "module" and sometimes not depending on context.

As a byproduct of this change I discovered a visibility bug for the
plplot_widget extension module in include/pldll.h.in that would have
had an effect in the Windows case.  This bug was fixed (along with
dropping "module" and "MODULE) as follows:

-#if defined ( plplot_widgetmodule_EXPORTS )
-  #define PLDLLIMPEXP_PLPLOT_WIDGETMODULE    PLDLLEXPORT
-  #define PLDLLIMPEXP_PLPLOT_MODULE_DATA( type )    PLDLLEXPORT type
+#if defined ( plplot_widget_EXPORTS )
+  #define PLDLLIMPEXP_PLPLOT_WIDGET    PLDLLEXPORT
+  #define PLDLLIMPEXP_PLPLOT_WIDGET_DATA( type )    PLDLLEXPORT type
 #else
-  #define PLDLLIMPEXP_PLPLOT_MODULE          PLDLLIMPORT
-  #define PLDLLIMPEXP_PLPLOT_MODULE_DATA( type )    PLDLLIMPORT type
+  #define PLDLLIMPEXP_PLPLOT_WIDGET    PLDLLIMPORT
+  #define PLDLLIMPEXP_PLPLOT_WIDGET_DATA( type )    PLDLLIMPORT type
 #endif

The (typographical) bug was the old code sometimes used MODULE when it
should have consistently used WIDGET_MODULE.  The _DATA part of this
bug was inconsequential because that construct was never used in the
plplot_widget C code.  So the net effect was no visibility bug on
Linux (since PLDLLIMPORT is not #defined on that platform), but there
would have been an issue on Windows where PLDLLIMPORT _is_ #defined.
But that bug was not detected because nobody has tested the pytkdemo
application (which uses the plplot_widget extension module) on Windows
(yet).

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by building the test_pytkdemo target (which tests the
plplot_widget extension module), the test_diff_psc target (which tests
the _plplotc extension module), and the test_pyqt[45]_example targets
(which tests the _plplotc extension module).  These tests were
sucessful for Python 2, but for Python 3 the test_pytkdemo target
failed (as expected since the current hand-crafted interface file,
bindings/python/plplot_widget.c is not compatible with what is needed
to create Python 3 extensions) as did the test_pyqt[45]_example
targets. I cannot explain that latter result because sip is designed
to generate interface files that are compatible with both Python 2 and
Python 3 and inspection of those generated files confirms that.  So I
have called for testing of the test_pyqt[45]_example targets on other
platforms to help discover whether this is a general issue or
something specific to the Debian Jessie platform.